Version 1.5
git-svn-id: https://plugins.svn.wordpress.org/woocommerce-easy-booking-system/trunk@1135433 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/settings/class-wceb-settings.php
@@ -24,7 +24,7 @@ class WCEB_Settings {
 		                            'easy_booking_text_color' => '#000000'
 		                        );
 
-		    add_option('easy_booking_settings', $this->options);
+		    add_option( 'easy_booking_settings', $this->options );
 
 		}
 
@@ -32,16 +32,25 @@ class WCEB_Settings {
 
 			add_action( 'admin_menu', array($this, 'easy_booking_add_option_page'), 10 );
 			add_action( 'admin_init', array($this, 'easy_booking_admin_init') );
-			add_action( 'easy_booking_save_settings', array($this, 'easy_booking_generate_css'), 10, 1 );
 
 		}
 
 	}
 
 	public function easy_booking_add_option_page() {
-		add_menu_page( __('Easy Booking', 'easy_booking'), __('Easy Booking', 'easy_booking'), 'manage_options', 'easy-booking', '', 'dashicons-calendar-alt', 58 );
+		$hook = add_menu_page( __('Easy Booking', 'easy_booking'), __('Easy Booking', 'easy_booking'), 'manage_options', 'easy-booking', '', 'dashicons-calendar-alt', 58 );
 		$option_page = add_submenu_page( 'easy-booking', __('Settings', 'easy_booking'), __('Settings', 'easy_booking'), 'manage_options', 'easy-booking', array($this, 'easy_booking_option_page') );
+
+		add_action( 'load-'. $hook, array($this, 'easy_booking_settings_save') );
 		add_action( 'admin_print_scripts-'. $option_page, array($this, 'easy_booking_load_admin_scripts') );
+	}
+
+	public function easy_booking_settings_save() {
+
+	  	if ( isset( $_GET['settings-updated'] ) && $_GET['settings-updated'] ) {
+	  		$this->easy_booking_generate_css();
+	   	}
+
 	}
 
 	public function easy_booking_load_admin_scripts() {
@@ -50,18 +59,31 @@ class WCEB_Settings {
 	}
 
 	// Generate static css file
-	public function easy_booking_generate_css( $data ) {
+	public function easy_booking_generate_css() {
 		$plugin_dir = plugin_dir_path( WCEB_PLUGIN_FILE ); // Shorten code, save 1 call
 
         $php_files = array(
-        	'default' => realpath ( $plugin_dir . 'assets/css/dev/default.min.css.php' ),
-        	'classic' => realpath ( $plugin_dir . 'assets/css/dev/classic.min.css.php' )
+        	'default' => realpath( $plugin_dir . 'assets/css/dev/default.min.css.php' ),
+        	'classic' => realpath( $plugin_dir . 'assets/css/dev/classic.min.css.php' )
         );
 
-        $css_files = array(
-        	'default' => realpath ( $plugin_dir . 'assets/css/default.min.css' ),
-        	'classic' => realpath ( $plugin_dir . 'assets/css/classic.min.css' )
-        );
+        if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			$blog_id = get_current_blog_id();
+
+			$default = str_replace('/', '\\', $plugin_dir . 'assets/css/default.' . $blog_id . '.min.css' ); // Replace backslashes with forwardslashes
+			$classic = str_replace('/', '\\', $plugin_dir . 'assets/css/classic.' . $blog_id . '.min.css' ); // Replace backslashes with forwardslashes
+
+			$css_files = array(
+	        	'default' => $default,
+	        	'classic' => $classic
+	        );
+
+		} else {
+			$css_files = array(
+	        	'default' => realpath( $plugin_dir . 'assets/css/default.min.css' ),
+	        	'classic' => realpath( $plugin_dir . 'assets/css/classic.min.css' )
+	        );
+		}
 
         if ( $php_files ) foreach ( $php_files as $theme => $php_file ) {
         	ob_start(); // Capture all output (output buffering)
@@ -70,8 +92,21 @@ class WCEB_Settings {
 	        
 	        $css = ob_get_clean(); // Get generated CSS (output buffering)
 
-	        if ( is_writable( $css_files[$theme] ) )
-	        	file_put_contents($css_files[$theme], $css); // Save it
+	        if ( file_exists( $css_files[$theme] ) ) {
+
+	        	if ( is_writable( $css_files[$theme] ) )
+	        		file_put_contents( $css_files[$theme], $css ); // Save it
+
+	        } else {
+
+		        fopen( $css_files[$theme], 'w' );
+
+		        if ( is_writable( $css_files[$theme] ) ) {
+		        	fwrite( $css_files[$theme], $css );
+		        	fclose( $css_files[$theme] );
+		        }
+
+	        }
 
         }
         
@@ -178,19 +213,74 @@ class WCEB_Settings {
 	public function easy_booking_option_page() {
 
 		?><div class="wrap">
-	
-			<?php screen_icon('generic'); ?>
 
-			<h2><?php _e('WooCommerce Easy Booking settings', 'easy_booking'); ?></h2>
+			<div id="wceb-settings-container">
 
-			<form method="post" action="options.php">
+				<div id="wceb-settings">
 
-				<?php settings_fields('easy_booking_settings'); ?>
-				<?php do_settings_sections('easy_booking_settings'); ?>
-				 
-				<?php submit_button(); ?>
+					<h2><?php _e('WooCommerce Easy Booking settings', 'easy_booking'); ?></h2>
 
-			</form>
+					<form method="post" action="options.php">
+
+						<?php settings_fields('easy_booking_settings'); ?>
+						<?php do_settings_sections('easy_booking_settings'); ?>
+						 
+						<?php submit_button(); ?>
+
+					</form>
+
+				</div>
+
+				<div id="wceb-sidebar">
+					<div class="easy-booking-notice">
+
+						<div class="easy-booking-notice__inner">
+							<h4><?php _e('For any issue or suggestion, check the FAQ or the support forum.', 'easy_booking') ?></h4>
+							<a href="http://herownsweetway.com/product/woocommerce-easy-booking/" class="button easy-booking-button" target="_blank"><?php _e('FAQ', 'easy_booking'); ?></a>
+							<a href="http://herownsweetway.com/support/woocommerce-easy-booking/" class="button easy-booking-button" target="_blank"><?php _e('Support', 'easy_booking'); ?></a>
+						</div>
+
+						<?php $active_plugins = (array) get_option( 'active_plugins', array() );
+
+				        if ( is_multisite() ) {
+				            $active_plugins = array_merge( $active_plugins, get_site_option( 'active_sitewide_plugins', array() ) );
+				        }
+				        
+				        if ( ! ( array_key_exists( 'easy-booking-availability-check/easy-booking-availability-check.php', $active_plugins ) || in_array( 'easy-booking-availability-check/easy-booking-availability-check.php', $active_plugins ) ) ) { ?>
+							<div class="easy-booking-notice__inner">
+								<h4><?php _e('Easy Booking : Availability Check', 'easy_booking'); ?></h4>
+								<p>
+									<?php _e( 'This add-on will allow you to manage stocks and availabilties when renting or booking your products.', 'easy_booking' ); ?>
+								</p>
+								<ul>
+									<li><?php _e('Keeps track of every booked date and quantity booked for each product (keeps only future dates).', 'easy_booking'); ?></li>
+									<li><?php _e('Updates availabilities when an order is made, modified or cancelled.', 'easy_booking'); ?></li>
+									<li><?php _e('Prevents users to order unavailable quantity or out-of-stock products.', 'easy_booking'); ?></li>
+									<li><?php _e('Allows you to easily check which dates and quantities are available in the administration panel.', 'easy_booking'); ?></li>
+								</ul>
+								<a href="http://herownsweetway.com/product/easy-booking-availability-check/" class="button button-hero easy-booking-button easy-booking-button--large" target="_blank"><?php _e( 'Learn more', 'easy_booking' ); ?></a>
+							</div>
+						<?php } ?>
+						<?php if ( ! ( array_key_exists( 'easy-booking-duration-discounts/easy-booking-duration-discounts.php', $active_plugins ) || in_array( 'easy-booking-duration-discounts/easy-booking-duration-discounts.php', $active_plugins ) ) ) { ?>
+							<div class="easy-booking-notice__inner">
+								<h4><?php _e('Easy Booking : Duration Discounts', 'easy_booking'); ?></h4>
+								<p>
+									<?php _e( 'This add-on will allow you to add discounts to your products depending on the duration booked by your clients.', 'easy_booking' ); ?>
+								</p>
+								<ul>
+									<li><?php _e('Choose custom discount amounts.', 'easy_booking'); ?></li>
+									<li><?php _e('Choose between "Product % discount", "Product discount", "Total % discount" or "Total discount".', 'easy_booking'); ?></li>
+									<li><?php _e('Define booked duration to add your discount (E.g : from 10 to 20 days).', 'easy_booking'); ?></li>
+									<li><?php _e('Add as many discounts as you want per product or variation.', 'easy_booking'); ?></li>
+								</ul>
+								<p><?php _e('The plugin then calculates the new product price when the client orders, depending on the discounts set.', 'easy_booking'); ?></p>
+								<a href="http://herownsweetway.com/product/easy-booking-duration-discounts/" class="button button-hero easy-booking-button easy-booking-button--large" target="_blank"><?php _e( 'Learn more', 'easy_booking' ); ?></a>
+							</div>
+						<?php } ?>
+					</div>
+				</div>
+
+			</div>
 
 		</div>
 		<?php
